refactor: separate files and rename
